chore(collection): restructure collection
