+++ app/movies/[id]/page.tsx
@@ -1,7 +1,5 @@
 import Fade from "../../../components/elements/fade";
 import Billboard from "../../../components/sections/billboard";
-import Browse from "../../../components/sections/browse";
-import Explore from "../../../components/sections/collection";
 import Collection from "../../../components/sections/collection";
 import Showcase from "../../../components/sections/showcase";
 import Content from "../../../components/wrappers/content";
@@ -29,14 +27,8 @@ const Page = async ({ params: { id } }: any) => {
         <Showcase media={spotlightMedia} isMediaSelected={true} />
         <div className="content relative bg-background-dark">
           <Fade />
-          <Explore>
-            <Collection title="More like this" medias={similarMedias} />
-          </Explore>
-          <Browse
-            title="Popular Movies"
-            variant="primary"
-            medias={popularMovies}
-          />
+          <Collection.Portrait title="More like this" medias={similarMedias} />
+          <Collection.Landscape title="Popular Movies" medias={popularMovies} />
         </div>
       </Content>
     </>
